+++ blocklets/ai-runtime/api/src/routes/ai.ts
@@ -1,7 +1,6 @@
 import { ReadableStream } from 'stream/web';
 
 import { getAgent, getMemoryVariables, getProject } from '@api/libs/agent';
-import { getProjectFromAIStudio } from '@api/libs/ai-studio';
 import { uploadImageToImageBin } from '@api/libs/image-bin';
 import logger from '@api/libs/logger';
 import ExecutionCache from '@api/store/models/execution-cache';
@@ -32,12 +31,14 @@ import { Op } from 'sequelize';
 const router = Router();
 
 const callInputSchema = Joi.object<{
+  entryAid?: string;
   aid: string;
   working?: boolean;
   sessionId?: string;
   inputs?: { [key: string]: any };
   debug?: boolean;
 }>({
+  entryAid: Joi.string().empty(['', null]),
   aid: Joi.string().required(),
   working: Joi.boolean().empty(['', null]).default(false),
   sessionId: Joi.string().empty(['', null]),
@@ -50,13 +51,16 @@ const callInputSchema = Joi.object<{
 const checkProjectRequestLimit = async ({
   userId,
   role,
+  blockletDid,
   projectId,
 }: {
   userId: string;
   role?: string;
+  blockletDid?: string;
   projectId: string;
 }) => {
-  const project = await getProjectFromAIStudio({ projectId });
+  const project = await getProject({ blockletDid, projectId, rejectOnEmpty: true });
+
   // 不限制自己创建的项目
   if (project.createdBy && project.createdBy === userId) {
     return;
@@ -178,7 +182,15 @@ router.post('/call', user(), compression(), async (req, res) => {
 
   const taskId = nextTaskId();
 
-  const sessionId = input.sessionId ?? (await Session.create({ userId, projectId, agentId })).id;
+  const sessionId =
+    input.sessionId ??
+    (
+      await Session.create({
+        userId,
+        projectId,
+        agentId: input.entryAid ? parseIdentity(input.entryAid, { rejectWhenError: true }).agentId : agentId,
+      })
+    ).id;
 
   const history = await History.create({
     userId,
@@ -253,7 +265,7 @@ router.post('/call', user(), compression(), async (req, res) => {
   if (stream) emit({ type: AssistantResponseType.CHUNK, taskId, assistantId: agent.id, delta: {} });
 
   try {
-    await checkProjectRequestLimit({ userId, role: req.user?.role, projectId });
+    await checkProjectRequestLimit({ userId, role: req.user?.role, blockletDid, projectId });
 
     emit({
       type: AssistantResponseType.INPUT_PARAMETER,